This PR adds `SubKnotVector` and methods of `Base.view` for performance.