@hyrodium 2023-06-10
The performance on `issubset` is much imprroved.

**Before this PR**
```julia
julia> using BasicBSpline, BenchmarkTools

julia> P = BSplineSpace{0}(knotvector"  1 2 1 1")
BSplineSpace{0, Int64, KnotVector{Int64}}(KnotVector([3, 5, 5, 7, 9]))

julia> P′= BSplineSpace{2}(knotvector"1 3 4 3 3 1")
BSplineSpace{2, Int64, KnotVector{Int64}}(KnotVector([1, 3, 3, 3, 5, 5, 5, 5, 7, 7, 7, 9, 9, 9, 11]))

julia> P ⊆ P′
true

julia> @benchmark P ⊆ P′
BenchmarkTools.Trial: 10000 samples with 201 evaluations.
 Range (min … max):  384.657 ns …  14.072 μs  ┊ GC (min … max): 0.00% … 93.94%
 Time  (median):     420.149 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   481.209 ns ± 732.955 ns  ┊ GC (mean ± σ):  8.62% ±  5.50%

     ▂▄▆▇██▇▆▅▄▃▂▂▁▁▁ ▂▃▄▅▅▅▄▃▂▁                                ▃
  ▂▅▇█████████████████████████████▇▇▇▇▇██████▇▇▇▆▅▇▆▆▆▆▇▆▅▄▅▆▄▅ █
  385 ns        Histogram: log(frequency) by time        625 ns <

 Memory estimate: 832 bytes, allocs estimate: 8.
```

**After this PR**
```julia
julia> using BasicBSpline, BenchmarkTools

julia> P = BSplineSpace{0}(knotvector"  1 2 1 1")
BSplineSpace{0, Int64, KnotVector{Int64}}(KnotVector([3, 5, 5, 7, 9]))

julia> P′= BSplineSpace{2}(knotvector"1 3 4 3 3 1")
BSplineSpace{2, Int64, KnotVector{Int64}}(KnotVector([1, 3, 3, 3, 5, 5, 5, 5, 7, 7, 7, 9, 9, 9, 11]))

julia> P ⊆ P′
true

julia> @benchmark P ⊆ P′
BenchmarkTools.Trial: 10000 samples with 994 evaluations.
 Range (min … max):  31.729 ns … 175.514 ns  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     33.020 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   36.166 ns ±   7.525 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▂█▇▁ ▄▇ ▁▂▂▄▁                  ▁▂▁                           ▂
  ████▄██▆█████▇▆▇▆▅▆▆▅▅▃▄▄▄▄▅▅▅▆████▆▆▇▇▇▇▇▇▇▇▇▇█▇▅▇▇▇▆▅▆▆▅▆▅ █
  31.7 ns       Histogram: log(frequency) by time      63.8 ns <

 Memory estimate: 0 bytes, allocs estimate: 0.
```